added serverless framework and files to deploy to serverless but encountered issue with socket-io connection where it must use an http server connection
